blockchain stateless refactor
